+++ src/hooks/randomDefense/useRandomDefenseCreateMenu.ts
@@ -1,6 +1,6 @@
 import { useState, useRef } from 'react';
 import { isRandomDefenseFormData } from '~types/typeGuards';
-import { validateRandomDefenseFormData } from '~domains/randomDefense/validateRandomDefenseFormData';
+import { validateRandomDefenseFormData } from '~domains/randomDefense/randomDefenseFormDataValidator';
 import type { ChangeEventHandler, MouseEventHandler } from 'react';
 import type {
   RandomDefenseFormData,
